requirements.txt
torch
transformers
pillow
paddleocr